#5 added redirect to the invite page after click event
