+++ db.json
@@ -20,6 +20,121 @@
       "date": "1540423525710",
       "name": "Rancho San Joaquin Golf Course",
       "id": 4
+    },
+    {
+      "date": "1540490368685",
+      "name": "UGP",
+      "id": 5
+    },
+    {
+      "date": "1540490390575",
+      "name": "UGP",
+      "id": 6
+    },
+    {
+      "date": "1540490982714",
+      "name": "UGP",
+      "id": 7
+    },
+    {
+      "date": "1540491025517",
+      "name": "UGP",
+      "id": 8
+    },
+    {
+      "date": "1540491168533",
+      "name": "UGP",
+      "id": 9
+    },
+    {
+      "date": "1540491245959",
+      "name": "UGP",
+      "id": 10
+    },
+    {
+      "date": "1540491254536",
+      "name": "OCPGI Orange County Par Golf Institute",
+      "id": 11
+    },
+    {
+      "date": "1540491879817",
+      "name": "Electric Cart Warehouse",
+      "id": 12
+    },
+    {
+      "date": "1540491954277",
+      "name": "Newport Beach Golf Course",
+      "id": 13
+    },
+    {
+      "date": "1540492517587",
+      "name": "UGP",
+      "id": 14
+    },
+    {
+      "date": "1540496819957",
+      "name": "UGP",
+      "id": 15
+    },
+    {
+      "date": "1540496880387",
+      "name": "UGP",
+      "id": 16
+    },
+    {
+      "date": "1540496912120",
+      "name": "UGP",
+      "id": 17
+    },
+    {
+      "date": "1540497015659",
+      "name": "Electric Cart Warehouse",
+      "id": 18
+    },
+    {
+      "date": "1540497310024",
+      "name": "UGP",
+      "id": 19
+    },
+    {
+      "date": "1540497704374",
+      "name": "OCPGI Orange County Par Golf Institute",
+      "id": 20
+    },
+    {
+      "date": "1540498100271",
+      "name": "Newport Beach Golf Course",
+      "id": 21
+    },
+    {
+      "date": "1540498329871",
+      "name": "OCPGI Orange County Par Golf Institute",
+      "id": 22
+    },
+    {
+      "date": "1540498512923",
+      "name": "OCPGI Orange County Par Golf Institute",
+      "id": 23
+    },
+    {
+      "date": "1540498567523",
+      "name": "Newport Beach Golf Course",
+      "id": 24
+    },
+    {
+      "date": "1540498644907",
+      "name": "Newport Beach Golf Course",
+      "id": 25
+    },
+    {
+      "date": "1540498689430",
+      "name": "The Club Fix",
+      "id": 26
+    },
+    {
+      "date": "1540498723194",
+      "name": "Rancho San Joaquin Golf Course",
+      "id": 27
     }
   ]
 }
