+++ src/container/invite.js
@@ -1,15 +1,27 @@
 import React, { Component } from 'react'
+// import {
+//   Typography,
+//   TextField,
+//   List,
+//   ListItem,
+//   Dialog,
+//   DialogActions,
+//   DialogTitle,
+//   Slide,
+//   Button } from '@material-ui/core'
+// import { withStyles } from '@material-ui/core/styles'
 
 class Invite extends Component {
-  // constructor() {
-  //   super()
+  // constructor(props) {
+  //   super(props)
   // }
 
   render() {
-
+    const { courseName } = this.props.match.params
+    console.log(courseName)
     return (
       <div>
-        I am here!!!
+
       </div>
     )
   }
